+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - `Upcast` for safely casting up a type hierarchy
 - `SetLens`, lenses operating on `Set`s
+- `ToArray`, for converting `Iterable<A>` to `A[]` 
 
 ## [3.0.0] - 2018-05-04
 ### Changed